# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: University Course, Fee, Intake & Requirement Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: universityId
          in: query
          schema:
            type: ["integer", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: degreeLevel
          in: query
          schema:
            type: ["string", "null"]
        - name: studyMode
          in: query
          schema:
            type: ["string", "null"]
        - name: country
          in: query
          schema:
            type: ["string", "null"]
        - name: intakeMonth
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course with full details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /courses/{courseId}/intakes:
    get:
      operationId: listCourseIntakes
      tags: [intakes]
      summary: List intakes for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of intakes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Intake"
    post:
      operationId: createCourseIntake
      tags: [intakes]
      summary: Add an intake to a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIntakeBody"
      responses:
        "201":
          description: Created intake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Intake"
  /intakes/{id}:
    patch:
      operationId: updateIntake
      tags: [intakes]
      summary: Update an intake
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIntakeBody"
      responses:
        "200":
          description: Updated intake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Intake"
    delete:
      operationId: deleteIntake
      tags: [intakes]
      summary: Delete an intake
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /courses/{courseId}/english-requirements:
    get:
      operationId: listCourseEnglishRequirements
      tags: [english-requirements]
      summary: List English requirements for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of English requirements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnglishRequirement"
    post:
      operationId: createCourseEnglishRequirement
      tags: [english-requirements]
      summary: Add English requirement to a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnglishRequirementBody"
      responses:
        "201":
          description: Created English requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnglishRequirement"
  /english-requirements/{id}:
    patch:
      operationId: updateEnglishRequirement
      tags: [english-requirements]
      summary: Update an English requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnglishRequirementBody"
      responses:
        "200":
          description: Updated English requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnglishRequirement"
    delete:
      operationId: deleteEnglishRequirement
      tags: [english-requirements]
      summary: Delete an English requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /bulk/courses/download:
    get:
      operationId: downloadCourses
      tags: [bulk]
      summary: Download all courses as CSV
      parameters:
        - name: universityId
          in: query
          schema:
            type: ["integer", "null"]
        - name: degreeLevel
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /bulk/courses/upload:
    post:
      operationId: uploadCourses
      tags: [bulk]
      summary: Upload courses via CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUploadBody"
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUploadResult"
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        universityId:
          type: integer
        universityName:
          type: ["string", "null"]
        name:
          type: string
        category:
          type: ["string", "null"]
        subCategory:
          type: ["string", "null"]
        courseWebsite:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        durationTerm:
          type: ["string", "null"]
        studyMode:
          type: ["string", "null"]
        degreeLevel:
          type: ["string", "null"]
        studyLoad:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        courseStructure:
          type: ["string", "null"]
        careerOutcomes:
          type: ["string", "null"]
        otherTest:
          type: ["string", "null"]
        otherTestScore:
          type: ["string", "null"]
        otherRequirement:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - universityId
        - name
        - status
        - createdAt
        - updatedAt
    CourseDetail:
      type: object
      properties:
        id:
          type: integer
        universityId:
          type: integer
        universityName:
          type: ["string", "null"]
        name:
          type: string
        category:
          type: ["string", "null"]
        subCategory:
          type: ["string", "null"]
        courseWebsite:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        durationTerm:
          type: ["string", "null"]
        studyMode:
          type: ["string", "null"]
        degreeLevel:
          type: ["string", "null"]
        studyLoad:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        courseStructure:
          type: ["string", "null"]
        careerOutcomes:
          type: ["string", "null"]
        otherTest:
          type: ["string", "null"]
        otherTestScore:
          type: ["string", "null"]
        otherRequirement:
          type: ["string", "null"]
        status:
          type: string
        intakes:
          type: array
          items:
            $ref: "#/components/schemas/Intake"
        fees:
          type: array
          items:
            $ref: "#/components/schemas/Fee"
        englishRequirements:
          type: array
          items:
            $ref: "#/components/schemas/EnglishRequirement"
        academicRequirements:
          type: array
          items:
            $ref: "#/components/schemas/AcademicRequirement"
        scholarships:
          type: array
          items:
            $ref: "#/components/schemas/Scholarship"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - universityId
        - name
        - status
        - intakes
        - fees
        - englishRequirements
        - academicRequirements
        - scholarships
        - createdAt
        - updatedAt
    CourseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateCourseBody:
      type: object
      properties:
        universityId:
          type: integer
        name:
          type: string
        category:
          type: ["string", "null"]
        subCategory:
          type: ["string", "null"]
        courseWebsite:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        durationTerm:
          type: ["string", "null"]
        studyMode:
          type: ["string", "null"]
        degreeLevel:
          type: ["string", "null"]
        studyLoad:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        courseStructure:
          type: ["string", "null"]
        careerOutcomes:
          type: ["string", "null"]
        otherTest:
          type: ["string", "null"]
        otherTestScore:
          type: ["string", "null"]
        otherRequirement:
          type: ["string", "null"]
      required:
        - universityId
        - name
    UpdateCourseBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        subCategory:
          type: ["string", "null"]
        courseWebsite:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        durationTerm:
          type: ["string", "null"]
        studyMode:
          type: ["string", "null"]
        degreeLevel:
          type: ["string", "null"]
        studyLoad:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        courseStructure:
          type: ["string", "null"]
        careerOutcomes:
          type: ["string", "null"]
        otherTest:
          type: ["string", "null"]
        otherTestScore:
          type: ["string", "null"]
        otherRequirement:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    Intake:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        intakeMonth:
          type: string
        intakeDay:
          type: ["integer", "null"]
        intakeYear:
          type: ["integer", "null"]
        isOpen:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - courseId
        - intakeMonth
        - isOpen
        - createdAt
    CreateIntakeBody:
      type: object
      properties:
        intakeMonth:
          type: string
        intakeDay:
          type: ["integer", "null"]
        intakeYear:
          type: ["integer", "null"]
        isOpen:
          type: boolean
      required:
        - intakeMonth
        - isOpen
    Fee:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        internationalFee:
          type: ["number", "null"]
        feeTerm:
          type: ["string", "null"]
        feeYear:
          type: ["integer", "null"]
        currency:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - courseId
        - createdAt
...
    EnglishRequirement:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        testType:
          type: string
        listening:
          type: ["number", "null"]
        speaking:
          type: ["number", "null"]
        writing:
          type: ["number", "null"]
        reading:
          type: ["number", "null"]
        overall:
          type: ["number", "null"]
        testName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - courseId
        - testType
        - createdAt
    CreateEnglishRequirementBody:
      type: object
      properties:
        testType:
          type: string
        listening:
          type: ["number", "null"]
        speaking:
          type: ["number", "null"]
        writing:
          type: ["number", "null"]
        reading:
          type: ["number", "null"]
        overall:
          type: ["number", "null"]
        testName:
          type: ["string", "null"]
      required:
        - testType
    AcademicRequirement:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        academicLevel:
          type: ["string", "null"]
        academicScore:
          type: ["number", "null"]
        scoreType:
          type: ["string", "null"]
        academicCountry:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - courseId
        - createdAt
...
    Scholarship:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        name:
          type: string
        details:
          type: ["string", "null"]
        eligibilityCriteria:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - courseId
        - name
        - createdAt
...
    BulkUploadBody:
      type: object
      properties:
        csvData:
          type: string
      required:
        - csvData
    BulkUploadResult:
      type: object
      properties:
        inserted:
          type: integer
        updated:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - inserted
        - updated
        - errors